chore: refactor location of auth related code
