+++ apps/platform/pkg/auth/redis-session-store.go
@@ -7,7 +7,6 @@ import (
 
 	"github.com/icza/session"
 	"github.com/thecloudmasters/uesio/pkg/cache"
-	"github.com/thecloudmasters/uesio/pkg/sess"
 )
 
 type RedisSessionStore struct {
@@ -16,7 +15,7 @@ type RedisSessionStore struct {
 
 func NewRedisSessionStore() session.Store {
 	return &RedisSessionStore{
-		cache.NewRedisCache[session.Session]("session").WithExpiration(sess.SessionLifetime).WithInitializer(session.NewSession),
+		cache.NewRedisCache[session.Session]("session").WithExpiration(SessionLifetime).WithInitializer(session.NewSession),
 	}
 }
