+++ apps/platform/pkg/sess/session.go
@@ -2,62 +2,16 @@ package sess
 
 import (
 	"context"
-	"net/http"
 	"sort"
-	"time"
-
-	"github.com/icza/session"
 
 	"github.com/thecloudmasters/uesio/pkg/meta"
 )
-
-const SessionLifetime = 12 * time.Hour
-
-func CreateBrowserSession(w http.ResponseWriter, r *http.Request, user *meta.User, site *meta.Site) session.Session {
-	sess := session.NewSessionOptions(&session.SessOptions{
-		CAttrs: map[string]any{
-			"Site":   site.GetFullName(),
-			"UserID": user.ID,
-		},
-		// TODO: Make Session timeout configurable by App/Site
-		// https://github.com/TheCloudMasters/uesio/issues/2643
-		Timeout: SessionLifetime,
-	})
-	browserSession := session.Get(r)
-	if browserSession != nil {
-		session.Remove(browserSession, w)
-	}
-	// Remove any previous set-cookie headers
-	// icza updates existing to "" and then its Add method
-	// appends so we end up with two Set-Cookie headers. This ensures
-	// we only have one and it's the current one
-	w.Header().Del("Set-Cookie")
-	session.Add(sess, w)
-	return sess
-}
-
-func GetSessionAttribute(browserSession session.Session, key string) string {
-	value, ok := browserSession.CAttr(key).(string)
-	if !ok {
-		return ""
-	}
-	return value
-}
 
 func New(ID string, user *meta.User, site *meta.Site) *Session {
 	return &Session{
 		ID:          ID,
 		siteSession: NewSiteSession(site, user),
 	}
-}
-
-func Login(w http.ResponseWriter, r *http.Request, user *meta.User, site *meta.Site) *Session {
-	return New(CreateBrowserSession(w, r, user, site).ID(), user, site)
-}
-
-func Logout(w http.ResponseWriter, r *http.Request, publicUser *meta.User, s *Session) *Session {
-	// Login as the public user - Login will logout the current user
-	return Login(w, r, publicUser, s.GetSiteSession().GetSite())
 }
 
 type WorkspaceSession struct {
@@ -328,15 +282,6 @@ func (s *Session) GetWorkspaceID() string {
 		return s.workspaceSession.GetID()
 	}
 	return ""
-}
-
-// IsExpired returns true if the browser session's last access time, plus the timeout duration,
-// is prior to the current timestamp.
-func IsExpired(browserSession session.Session) bool {
-	if browserSession == nil {
-		return true
-	}
-	return browserSession.Accessed().Add(browserSession.Timeout()).Before(time.Now())
 }
 
 // IsPublicUser returns true if the session is for a public user (i.e., not logged in).
